# Migrate from difference to similar