--- a/Cargo.toml
+++ b/Cargo.toml
@@ -55,7 +55,7 @@ codespan-reporting = "0.11"
 crossbeam = "0.8"
 # We use this for displaying diffs for dry runs of the `fix` subcommand, as
 # as in tests for printing easy to read errors
-difference = "2.0"
+similar = "1.3"
 # Logging utilities
 fern = "0.6"
 # We directly interact with git when doing index operations eg during fix