### Checklist

* [X] I have read the Contributor Guide
* [X] I have read and agree to the Code of Conduct
* [X] I have added a description of my changes and why I'd like them included in the section below

### Description of Changes

Since `difference` is unmaintained, replace it with `similar` exactly matching the previous behavior.

Also remove unencountered ban skips

---

I debated using `dissimilar` vs `similar`

I tried using `dissimilar` first since: it's half the size, twice the downloads, and the maintainer is already sponsored by embark studios
However due to the way it finds differences I needed to write quite a bit of logic to replicate the structure output by the existing `difference` changeset

If you have these inputs:

original:
```
abc
def
```

new:
```
abc
dgg
```

It gives the changes like so:

unchanged:
```
abc
d
```

deleted:
```
ef
```

inserted:
```
gg
```

so you then need to account for the shared `d` and just generally it was more code than you probably want to maintain for a simple diff in an info log people wont see on the default warn level

`similar` had a line by line diff which made it easy to recreate the `difference` behavior in a couple of extra lines

---

I used manual escape codes rather than `ansi_term` to keep the colors identical to `difference`

https://github.com/EmbarkStudios/cargo-deny/blob/88dda4a8c827af9b06ab2cc0e4bac0656dd2f7bd/src/cargo-deny/fix.rs#L127-L128

---

I saw in a PR swapping difference to similar they had speed regression issues but I found the speed to be about the same (slightly faster in a quite quick and unscientific test)

https://github.com/colin-kiegel/rust-pretty-assertions/pull/51

```
$ hyperfine --warmup 3 '../../target/debug/cargo-deny -L info fix --dry-run' './new/cargo-deny -L info fix --dry-run'
Benchmark #1: ../../target/debug/cargo-deny -L info fix --dry-run
  Time (mean ± σ):      1.428 s ±  0.073 s    [User: 1.142 s, System: 0.286 s]
  Range (min … max):    1.325 s …  1.551 s    10 runs

Benchmark #2: ./new/cargo-deny -L info fix --dry-run
  Time (mean ± σ):      1.387 s ±  0.063 s    [User: 1.105 s, System: 0.282 s]
  Range (min … max):    1.321 s …  1.510 s    10 runs

Summary
  './new/cargo-deny -L info fix --dry-run' ran
    1.03 ± 0.07 times faster than '../../target/debug/cargo-deny -L info fix --dry-run'
```

---

The binary size does increase by about 2M :disappointed: 

```
$ du ../../target/debug/cargo-deny
160672	../../target/debug/cargo-deny

$ du ./new/cargo-deny
162464	./new/cargo-deny
```

### Related Issues

https://github.com/EmbarkStudios/krates/pull/32